refactor(bootstrap): extract @workglow/bootstrap from duplicated registries
`bootstrapWorkglow()` and the vitest harness's `bootstrapTestRegistry()` each
made the same 14 default-registration calls in the same order. The two were
provably equivalent — every registration function defaults its registry
parameter to `globalServiceRegistry`, so the no-argument test variant and the
explicit-registry production variant did identical work.

Promote the shared logic to a new `@workglow/bootstrap` package sitting below
both consumers, depending only on util, storage, task-graph, knowledge-base,
ai, and mcp.

- `registerAllDefaults(registry = globalServiceRegistry)` is now the single
  definition, with the signature widened so both call styles are served.
- `bootstrapWorkglow()`, `createOrchestrationContext()`, `BootstrapOptions`,
  and `WorkglowContext` move alongside it.
- `packages/workglow/src/bootstrap.ts` becomes a re-export, so the
  `workglow/bootstrap` subpath and the root barrel keep working unchanged.
  `auto-bootstrap.ts` stays in the meta-package, which carries the tslog
  dependency its logger needs.
- `vitest.setup.ts` calls `registerAllDefaults()` directly, imported by source
  path because the isolated linker keeps workspace packages out of the repo
  root's node_modules.
